+++ packages/react/src/themes/teams/components/Toolbar/toolbarItemStyles.ts
@@ -3,14 +3,20 @@ import { ToolbarItemProps } from '../../../../components/Toolbar/ToolbarItem'
 import { ToolbarVariables } from './toolbarVariables'
 import getIconFillOrOutlineStyles from '../../getIconFillOrOutlineStyles'
 import { getColorScheme } from '../../colors'
+import getBorderFocusStyles from '../../getBorderFocusStyles'
 
 const toolbarItemStyles: ComponentSlotStylesPrepared<ToolbarItemProps, ToolbarVariables> = {
-  root: ({ props: p, variables: v }): ICSSInJSStyle => {
+  root: ({ props: p, variables: v, theme: { siteVariables } }): ICSSInJSStyle => {
     const colors = getColorScheme(v.colorScheme)
+    const { borderWidth } = siteVariables
+    const borderFocusStyles = getBorderFocusStyles({
+      siteVariables,
+    })
 
     return {
+      position: 'relative',
       backgroundColor: v.background,
-      borderWidth: v.borderWidth,
+      borderWidth,
       borderStyle: 'solid',
       borderColor: 'transparent',
       borderRadius: v.borderRadius,
@@ -20,9 +26,7 @@ const toolbarItemStyles: ComponentSlotStylesPrepared<ToolbarItemProps, ToolbarVa
       color: v.foreground || colors.foreground1,
       cursor: 'pointer',
 
-      ':focus': {
-        outline: 0,
-      },
+      ':focus': borderFocusStyles[':focus'],
 
       ...(p.active && {
         color: v.foregroundActive || colors.foregroundActive,
@@ -36,12 +40,7 @@ const toolbarItemStyles: ComponentSlotStylesPrepared<ToolbarItemProps, ToolbarVa
         ...getIconFillOrOutlineStyles({ outline: false }),
       },
 
-      ':focus-visible': {
-        color: v.foregroundFocus || colors.foregroundFocus,
-        backgroundColor: v.backgroundFocus || colors.backgroundFocus,
-        borderColor: v.borderFocus || colors.borderFocus,
-        ...getIconFillOrOutlineStyles({ outline: false }),
-      },
+      ':focus-visible': borderFocusStyles[':focus-visible'],
 
       ...(p.disabled && {
         color: v.foregroundDisabled || colors.foregroundDisabled1,